requirements.txt
chainlit
SpeechRecognition
PyAudio
pydub
numpy
openai
google-genai
python-dotenv
langchain
langchain-core
langchain-community
langchain-chroma
langchain-google-genai
PyPDF2
python-docx
pillow
scikit-learn
beautifulsoup4
html2text
duckduckgo-search
librosa
scipy
noisereduce
asyncpg
SQLAlchemy
aiohttp